+++ src/Filesystem/MountManager/MountManagerFactory.php
@@ -34,9 +34,13 @@ class MountManagerFactory implements FactoryInterface
             foreach ($config['filesystem']['adapters'] as $prefix => $adapter) {
                 $class = $adapter['class'];
                 $arguments = !empty($adapter['arguments']) ? $adapter['arguments'] : [];
+                $filesystemConfig = null;
                 try {
                     if ($arguments) {
                         if ($container instanceof ServiceManager) {
+                            if (!empty($arguments['config'])) {
+                                $filesystemConfig = $arguments['config']; unset($arguments['config']);
+                            }
                             $filesystemAdapter = $container->build($class, $arguments);
                         } else {
                             throw new Exception\LogicException(
@@ -49,8 +53,8 @@ class MountManagerFactory implements FactoryInterface
                 } catch (ServiceNotCreatedException $e) {
                     throw new Exception\RuntimeException("Error in filesystem/adapters/$prefix configuration", 0, $e);
                 }
-
-                $filesystems[$prefix] = new Filesystem($filesystemAdapter);
+print_r($filesystemConfig);
+                $filesystems[$prefix] = new Filesystem($filesystemAdapter, $filesystemConfig);
             }
         }