+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Conductor: Core
 - Removed composer requirements that were pushed to other repos
 - Added Crypt commands and model
 - Added posix PHP extension suggestion
+- Added support for setting config on filesystems
 
 # 0.1.0
 - Initial build copied over from devops tool